chore(*): Minor refactor. Move  common code to util and re-use in selector and header.
