+++ projects/igniteui-angular/src/lib/grids/pivot-grid/pivot-header-row.component.ts
@@ -281,29 +281,6 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent implem
     * @hidden
     * @internal
     */
-    public getAggregateList(val: IPivotValue): IPivotAggregator[] {
-        if (!val.aggregateList) {
-            let defaultAggr = this.getAggregatorsForValue(val);
-            const isDefault = defaultAggr.find(x => x.key === val.aggregate.key);
-            // resolve custom aggregations
-            if (!isDefault && this.grid.data[0][val.member] !== undefined) {
-                // if field exists, then we can apply default aggregations and add the custom one.
-                defaultAggr.unshift(val.aggregate);
-            } else if (!isDefault) {
-                // otherwise this is a custom aggregation that is not compatible
-                // with the defaults, since it operates on field that is not in the data
-                // leave only the custom one.
-                defaultAggr = [val.aggregate];
-            }
-            val.aggregateList = defaultAggr;
-        }
-        return val.aggregateList;
-    }
-
-    /**
-    * @hidden
-    * @internal
-    */
     public rowRemoved(event: IBaseChipEventArgs) {
         const row = this.grid.pivotConfiguration.rows.find(x => x.memberName === event.owner.id);
         this.grid.toggleDimension(row);
@@ -592,26 +569,10 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent implem
         return isColumn ? PivotDimensionType.Column : isRow ? PivotDimensionType.Row : PivotDimensionType.Filter;
     }
 
-    protected getAggregatorsForValue(value: IPivotValue): IPivotAggregator[] {
-        const dataType = value.dataType || this.grid.resolveDataTypes(this.grid.data[0][value.member]);
-        switch (dataType) {
-            case GridColumnDataType.Number:
-            case GridColumnDataType.Currency:
-                return IgxPivotNumericAggregate.aggregators();
-            case GridColumnDataType.Date:
-            case GridColumnDataType.DateTime:
-                return IgxPivotDateAggregate.aggregators();
-            case GridColumnDataType.Time:
-                return IgxPivotTimeAggregate.aggregators();
-            default:
-                return IgxPivotAggregate.aggregators();
-        }
-    }
-
     protected updateDropDown(value: IPivotValue, dropdown: IgxDropDownComponent, chip: IgxChipComponent) {
         this.value = value;
         dropdown.width = chip.nativeElement.clientWidth + 'px';
-        this.aggregateList = this.getAggregateList(value);
+        this.aggregateList =  PivotUtil.getAggregateList(value, this.grid);
         dropdown.open(this._subMenuOverlaySettings);
     }
 }
